pyproject.toml
[project]
name = "jpeg-tb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=11.3.0",
    "scipy>=1.17.1",
]
